pages(forgot-password.vue,reset-password.vue)のコンポーネント分割 #17
